+++ api/src/test/java/ca/bc/gov/educ/api/penrequest/controller/DocumentWebMvcTests.java
@@ -7,10 +7,10 @@ import ca.bc.gov.educ.api.penrequest.props.ApplicationProperties;
 import ca.bc.gov.educ.api.penrequest.repository.DocumentRepository;
 import ca.bc.gov.educ.api.penrequest.repository.DocumentTypeCodeTableRepository;
 import ca.bc.gov.educ.api.penrequest.repository.PenRequestRepository;
-import ca.bc.gov.educ.api.support.DocumentBuilder;
-import ca.bc.gov.educ.api.support.DocumentTypeCodeBuilder;
-import ca.bc.gov.educ.api.support.PenRequestBuilder;
-import ca.bc.gov.educ.api.support.WithMockOAuth2Scope;
+import ca.bc.gov.educ.api.penrequest.support.DocumentBuilder;
+import ca.bc.gov.educ.api.penrequest.support.DocumentTypeCodeBuilder;
+import ca.bc.gov.educ.api.penrequest.support.PenRequestBuilder;
+import ca.bc.gov.educ.api.penrequest.support.WithMockOAuth2Scope;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
