+++ api/src/test/java/ca/bc/gov/educ/api/penrequest/controller/PenRequestCommentsControllerTest.java
@@ -4,7 +4,7 @@ import ca.bc.gov.educ.api.penrequest.exception.RestExceptionHandler;
 import ca.bc.gov.educ.api.penrequest.model.PenRequestEntity;
 import ca.bc.gov.educ.api.penrequest.repository.PenRequestCommentRepository;
 import ca.bc.gov.educ.api.penrequest.repository.PenRequestRepository;
-import ca.bc.gov.educ.api.support.WithMockOAuth2Scope;
+import ca.bc.gov.educ.api.penrequest.support.WithMockOAuth2Scope;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.BeforeClass;
