+++ api/src/test/java/ca/bc/gov/educ/api/penrequest/service/DocumentServiceTests.java
@@ -6,9 +6,9 @@ import ca.bc.gov.educ.api.penrequest.model.PenRequestEntity;
 import ca.bc.gov.educ.api.penrequest.repository.DocumentRepository;
 import ca.bc.gov.educ.api.penrequest.repository.DocumentTypeCodeTableRepository;
 import ca.bc.gov.educ.api.penrequest.repository.PenRequestRepository;
-import ca.bc.gov.educ.api.support.DocumentBuilder;
-import ca.bc.gov.educ.api.support.DocumentTypeCodeBuilder;
-import ca.bc.gov.educ.api.support.PenRequestBuilder;
+import ca.bc.gov.educ.api.penrequest.support.DocumentBuilder;
+import ca.bc.gov.educ.api.penrequest.support.DocumentTypeCodeBuilder;
+import ca.bc.gov.educ.api.penrequest.support.PenRequestBuilder;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
