+++ api/src/test/java/ca/bc/gov/educ/api/penrequest/support/WithMockOAuth2Scope.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.educ.api.support;
+package ca.bc.gov.educ.api.penrequest.support;
 
 import org.springframework.security.test.context.support.WithSecurityContext;
